requirements.txt
streamlit==1.32.2
folium==0.16.0
streamlit-folium==0.19.0
pandas==2.2.1
geopandas==0.14.3
SQLAlchemy==2.0.28
GeoAlchemy2==0.14.6
requests==2.31.0
flake8==7.0.0
pytest==8.1.1
prefect==2.16.5
python-dotenv==1.0.1
tenacity==8.2.3
psycopg2-binary==2.9.9
